refactor(dropzone-image): simplify props, refactor
